fixed some mistakes

+++ libft/ft_strdup.c
@@ -10,9 +10,6 @@
 /*                                                                            */
 /* ************************************************************************** */
 
-#include <string.h>
-#include <stdio.h>
-#include <stdlib.h>
 #include "libft.h"
 
 char	*ft_strdup(const char *s1)
@@ -21,7 +18,7 @@ char	*ft_strdup(const char *s1)
 	char	*s2;
 
 	i = 0;
-	s2 = malloc((size_t)ft_strlen(s1)+ 1);
+	s2 = malloc(ft_strlen(s1)+ 1);
 	if (!s2)
 	{
 		return (NULL);
